fix: seed palace identity for local-memory-only deployments too
Seeding lived in MemPalaceRemoteService.start(), which only runs when
MEMPALACE_REMOTE_ENABLED is true; a local-only deployment (or a remote
startup failure) still woke agents identity-less. Move the template and
seeding into memPalaceClient (ensurePalaceIdentity) and call it from
index.ts whenever MEMPALACE_ENABLED is set, before the MCP starts.
