+++ src/services/memPalaceRemoteService.ts
@@ -12,33 +12,6 @@ import { spawnCollect } from "../utils/spawnCollect.js";
 import { buildRepoCheckoutPath } from "./gitWorkspaceService.js";
 
 const GENERATED_HEADER = "# Generated by Actuarius. Do not commit; this file routes repo memory to MemPalace.";
-/**
- * Seeded into the palace when no identity.txt exists, so agents waking up via
- * mempalace_wake_up receive baseline conduct rules. Written once; operator
- * edits on the persistent disk are never overwritten.
- */
-const IDENTITY_TEMPLATE = `# Identity — Actuarius agents
-
-You are an AI agent (Claude, Codex, Gemini, or OpenCode) executing requests
-for the Actuarius Discord bot. Requests arrive from Discord threads and run in
-isolated git worktrees on a branch created for the request.
-
-## Git rules
-- NEVER push directly to \`main\` or \`master\`. Work on the request branch you
-  were given; changes reach main through pull requests.
-- Do not force-push or rewrite published history.
-- Do not run destructive git commands (\`git reset --hard\`, \`git checkout --\`)
-  on work you did not create in this request.
-- Use \`gh\` for author-sensitive GitHub actions (pull requests, comments,
-  review replies) so the bot identity is used.
-
-## Memory rules
-- Repo knowledge belongs in the repo's wing (see mempalace.yaml in your
-  worktree); use the branch name as the room.
-- Verify facts about people, projects, or past events with mempalace search
-  or kg queries before asserting them. Never guess.
-`;
-
 const DEFAULT_ROOMS = [
   { name: "src", keywords: ["src", "source", "code", "implementation"] },
   { name: "tests", keywords: ["test", "tests", "spec", "fixture"] },
@@ -199,7 +172,6 @@ export class MemPalaceRemoteService {
       throw new MemPalaceRemoteConfigError("mempalace-cli binary not found at " + this.config.mempalaceCliPath);
     }
     await this.ensureToken();
-    await this.ensureIdentityFile();
     await this.configureKnownRepositories(existingRepos);
     await this.writeGlobalConfig();
     this.serverDesired = true;
@@ -210,23 +182,6 @@ export class MemPalaceRemoteService {
     this.serverDesired = false;
     this.clearServerRetry();
     await this.runServerOp(() => this.stopGuarded());
-  }
-
-  /**
-   * Seed the palace identity file agents receive from mempalace_wake_up.
-   * Only writes when the file is absent so operator edits persist; failure is
-   * non-fatal — agents just wake up without an identity, as before.
-   */
-  private async ensureIdentityFile(): Promise<void> {
-    const identityPath = join(this.homeDir, ".mempalace", "identity.txt");
-    if (await pathExists(identityPath)) return;
-    try {
-      await mkdir(dirname(identityPath), { recursive: true });
-      await writeFile(identityPath, IDENTITY_TEMPLATE, "utf8");
-      this.logger.info({ identityPath }, "Seeded MemPalace identity file");
-    } catch (error) {
-      this.logger.warn({ error, identityPath }, "Could not seed MemPalace identity file");
-    }
   }
 
   public async configureKnownRepositories(rows: RepoRow[]): Promise<void> {
